+++ tests/fixtures/example-fail.js
@@ -1,3 +1,7 @@
+/**
+ * SPDX-FileCopyrightText: 2023 Nextcloud GmbH and Nextcloud contributors
+ * SPDX-License-Identifier: AGPL-3.0-or-later
+ */
 //should fail because of missing space on this comment
 // and usage of `console` and using a semicolon
 console.log('some');